YMG/Yazilim-Muhendisliginde-Gelismeler-1-main/requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
psycopg2-binary
requests
prometheus_client
mcp
ollama
# Güvenlik için eklenenler:
passlib[bcrypt]
python-jose[cryptography]